requirements.txt
streamlit
openai
faiss-cpu
sentence-transformers
torch
Pillow
ftfy
regex
tqdm
open-clip-torch
python-dotenv
python-docx
pymupdf
